fix(geo): change timeZoneOffset to double in NotificareRegion

+++ notificare_geo/lib/src/models/notificare_region.dart
@@ -13,7 +13,7 @@ class NotificareRegion {
   final int? major;
   final double distance;
   final String timeZone;
-  final int timeZoneOffset;
+  final double timeZoneOffset;
 
   NotificareRegion({
     required this.id,
